# restapi/openapi-full.yaml
openapi: 3.0.0
info:
  title: WinCC OA REST API
  version: 1.0.0
  description: |
    REST API for WinCC OA GraphQL Server - provides HTTP endpoints for all WinCC OA operations including:
    - Data point management (CRUD, value operations)
    - Data point type management
    - Tag queries with metadata
    - Alert handling
    - CNS (Central Navigation Service) operations
    - System information and redundancy
    - OPC UA configuration
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /restapi/auth/login
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Missing required parameter
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ElementType:
      type: string
      enum:
        - BOOL
        - UINT8
        - INT32
        - INT64
        - FLOAT
        - DOUBLE
        - BIT
        - BIT32
        - BIT64
        - STRING
        - TIME
        - DPID
        - LANGSTRING
        - BLOB
        - MIXED
        - DYN_BOOL
        - DYN_UINT8
        - DYN_INT32
        - DYN_INT64
        - DYN_FLOAT
        - DYN_DOUBLE
        - DYN_BIT
        - DYN_BIT32
        - DYN_BIT64
        - DYN_STRING
        - DYN_TIME
        - DYN_DPID
        - DYN_LANGSTRING
        - DYN_BLOB
...
    DpTypeNode:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          example: value
        type:
          $ref: '#/components/schemas/ElementType'
        refName:
          type: string
          example: RefType
        children:
          type: array
          items:
            $ref: '#/components/schemas/DpTypeNode'
        newName:
          type: string
          example: newValue
    Tag:
      type: object
      properties:
        name:
          type: string
          example: ExampleDP_1.value
        value:
          description: Current tag value (any type)
          example: 42.5
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        status:
          type: object
          example: { "_online": { "_value": true } }
...
    AlertTime:
      type: object
      required: [time, count, dpe]
      properties:
        time:
          type: string
          format: date-time
        count:
          type: integer
        dpe:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Missing or invalid authentication token
    Forbidden:
      description: Insufficient permissions (admin role required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: Admin role required
paths:
...
  /restapi/datapoints:
    get:
      tags: [Data Points]
      summary: List data points
      description: Search for data points by pattern
      security:
        - bearerAuth: []
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
          description: Search pattern (wildcards supported)
          example: "ExampleDP_*"
        - name: dpType
          in: query
          schema:
            type: string
          description: Filter by data point type
        - name: ignoreCase
          in: query
          schema:
            type: boolean
          description: Case-insensitive search
      responses:
        '200':
          description: List of data points
          content:
            application/json:
              schema:
                type: object
                properties:
                  datapoints:
                    type: array
                    items:
                      type: string
                    example: ["ExampleDP_Arg1.", "ExampleDP_Arg2."]
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Data Points]
      summary: Create data point
      description: Create a new data point (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dpeName, dpType]
              properties:
                dpeName:
                  type: string
                  example: "MyDataPoint."
                dpType:
                  type: string
                  example: "ExampleDP_Arg"
                systemId:
                  type: integer
                  example: 1
                dpId:
                  type: integer
                  example: 100
      responses:
        '200':
          description: Data point created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Data Points - Value Operations
...
  /restapi/query:
    post:
      tags: [Data Points]
      summary: Execute SQL-like query on data points
      description: |
        Retrieves attribute values with the help of SQL statements.
        Returns a table-like structure where:
        - [0][0] is empty
        - [0][1..n] are column headers (attribute names)
        - [1..n][0] are line names (data point names)
        - [1..n][1..n] are the values

        Example query: `SELECT '_original.._value' FROM 'ExampleDP_Arg*'`

        This is equivalent to the WinCC OA `dpQuery()` function.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: SQL statement for querying data points
                  example: "SELECT '_original.._value' FROM 'ExampleDP_Arg*'"
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    description: Table-like result structure
                    items:
                      type: array
                      items:
                        description: Cell value (any type)
                    example:
                      - ["", ":_original.._value"]
                      - ["System1:ExampleDP_Arg1.", 2.43]
                      - ["System1:ExampleDP_Arg2.", 5.76]
        '400':
          description: Bad request - query parameter missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Data Point Types
  /restapi/datapoint-types:
    get:
      tags: [Data Point Types]
      summary: List data point types
      security:
        - bearerAuth: []
      parameters:
        - name: pattern
          in: query
          schema:
            type: string
        - name: systemId
          in: query
          schema:
            type: integer
        - name: includeEmpty
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of types
          content:
            application/json:
              schema:
                type: object
                properties:
                  dpTypes:
                    type: array
                    items:
                      type: string

    post:
      tags: [Data Point Types]
      summary: Create data point type
      description: Create a new data point type (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startNode]
              properties:
                startNode:
                  $ref: '#/components/schemas/DpTypeNode'
      responses:
        '200':
          description: Type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /restapi/tags:
    get:
      tags: [Tags]
      summary: Get tags
      description: Get multiple tags with value, timestamp, and status
      security:
        - bearerAuth: []
      parameters:
        - name: dpeNames
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated list of data point element names
          example: "ExampleDP_1.value,ExampleDP_2.value"
      responses:
        '200':
          description: Tags retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
...
  /restapi/alerts:
    get:
      tags: [Alerts]
      summary: Get alerts
      description: Get alert attributes
      security:
        - bearerAuth: []
      parameters:
        - name: alertsTime
          in: query
          required: true
          schema:
            type: string
          description: JSON-encoded array of alert times
        - name: dpeNames
          in: query
          required: true
          schema:
            type: string
          description: JSON-encoded array of DPE names
        - name: alertCount
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Alert values
          content:
            application/json:
              schema:
                type: object
                properties:
                  values:
                    type: array

    put:
      tags: [Alerts]
      summary: Set alert attributes
      description: Set alert attributes (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [alerts, values]
              properties:
                alerts:
                  type: array
                  items:
                    $ref: '#/components/schemas/AlertTime'
                values:
                  type: array
      responses:
        '200':
          description: Alerts set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...